Updated Login Logic with react-router-dom 6.4 and Replaced UseAuthentication with Login
Implemented a change in the login logic in the Login file to utilize the functionalities of react-router-dom 6.4. This update improves the login process by leveraging the latest features provided by the updated react-router-dom version. Additionally, replaced the UseAuthentication component with the Login functionality, resulting in a more streamlined and intuitive authentication flow.